+++ araucaria/linalg/__init__.py
@@ -14,5 +14,7 @@ perform linear algebra operations:
      - Computes the condition number of a collection.
    * - :func:`~araucaria.linalg.la.matrix_rank`
      - Computes the matrix rank of a collection.
+   * - :func:`~araucaria.linalg.la.imd`
+     - Computes the interpolative matrix decomposition of a collection.
 """
-from .la import cond_num, matrix_rank
+from .la import cond_num, matrix_rank, imd