+++ docs/source/changelog.rst
@@ -3,8 +3,9 @@ Changelog
 
 ver 0.1.11
 ----------
-- :mod:`~.linalg` module added: linar algebra operations on collections.
+- :mod:`~.stats.signal` submodule added: functions to filter and analyze univariate data.
+- :mod:`~.linalg` module added: linear algebra operations on collections.
 - :mod:`~.xrdb` module added: basic support for x-ray databases.
-- added :func:`~.utils.read_fdicts` function (utilities module).
-- fixed bug in :func:`~.xasutils.get_mapped_data` function (Morgane Desmau).
+- added :func:`~.utils.read_fdicts` function (:mod:`~araucaria.utils` module).
+- fixed bug in derivative computation of :func:`~.xasutils.get_mapped_data` function (thanks to M. Desmau).
 - updated tutorials.